backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
email-validator==2.0.0.post2
requests==2.29.0
pyjwt==2.6.0
psycopg2-binary==2.9.6